backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
jinja2
python-dotenv
psycopg2-binary
requests